src/test/java/com/imooc/Service/imp/CategoryServiceImpTest.java
package com.imooc.Service.imp;

import com.imooc.dataobject.ProductCategory;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.Arrays;
import java.util.List;

import static org.junit.Assert.*;

@RunWith(SpringRunner.class)
@SpringBootTest
public class CategoryServiceImpTest {

    @Autowired
    CategoryServiceImp categoryService;

    @Test
    public void findOne() {
        ProductCategory productCategory = categoryService.findOne(1);
        Assert.assertEquals(new Integer(1),productCategory.getCategoryId());
    }

    @Test
    public void findAll() {
        List<ProductCategory> list = categoryService.findAll();
        Assert.assertNotEquals(0,list.size());
    }

    @Test
    public void save() {
        ProductCategory productCategory = new ProductCategory("测试", 100);
        ProductCategory result = categoryService.save(productCategory);
        Assert.assertNotEquals(null, result);
    }

    @Test
    public void findBycategoryType() {
        List<ProductCategory> list = categoryService.findBycategoryType(Arrays.asList(3));
        Assert.assertNotEquals(0,list.size());
    }
}